scheduler/requirements.txt
google-cloud-firestore
firebase-admin
google-cloud-pubsub